update from upstream to get the 64 bit counter changes

+++ src/runtime_src/driver/include/xcl_app_debug.h
@@ -43,11 +43,9 @@ extern "C" {
 #endif
 
 /************************ SPM Debug Counters ********************************/
-//debug is only interested in 4 metric counters: wb,wt,rb,rt,outstanding,lwa,lwd,lra,lrd
 #define XSPM_DEBUG_SAMPLE_COUNTERS_PER_SLOT     9
 
 /************************ SSPM Debug Counters ********************************/
-//debug is interested in all of the counter values
 #define XSSPM_DEBUG_SAMPLE_COUNTERS_PER_SLOT    5
 
 /*
@@ -78,16 +76,16 @@ enum xclDebugReadType {
 
 /* Debug counter results */
 typedef struct {
-  unsigned int           WriteBytes     [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           WriteTranx     [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           ReadBytes      [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           ReadTranx      [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int WriteBytes     [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int WriteTranx     [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int ReadBytes      [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int ReadTranx      [XSPM_MAX_NUMBER_SLOTS];
 
-  unsigned int           OutStandCnts   [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           LastWriteAddr  [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           LastWriteData  [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           LastReadAddr   [XSPM_MAX_NUMBER_SLOTS];
-  unsigned int           LastReadData   [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int OutStandCnts   [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int LastWriteAddr  [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int LastWriteData  [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int LastReadAddr   [XSPM_MAX_NUMBER_SLOTS];
+  unsigned long long int LastReadData   [XSPM_MAX_NUMBER_SLOTS];
   unsigned int           NumSlots;
   char                   DevUserName    [256];
 } xclDebugCountersResults;
